# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertKind": {
        "description": "What raised an alert.",
        "enum": [
          "identity_match",
          "rule"
        ],
        "title": "AlertKind",
        "type": "string"
      },
      "AlertPageResponse": {
        "description": "One page of alerts, newest first.",
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "title": "Alerts",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "alerts",
          "next_cursor"
        ],
        "title": "AlertPageResponse",
        "type": "object"
      },
...
      "AlertSummaryResponse": {
        "description": "Totals of the alerts that match a filter.",
        "properties": {
          "counts": {
            "items": {
              "$ref": "#/components/schemas/AlertCountResponse"
            },
            "title": "Counts",
            "type": "array"
          },
          "daily_counts": {
            "items": {
              "$ref": "#/components/schemas/DailyAlertCountResponse"
            },
            "title": "Daily Counts",
            "type": "array"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "unacknowledged_count": {
            "title": "Unacknowledged Count",
            "type": "integer"
          }
        },
        "required": [
          "total_count",
          "unacknowledged_count",
          "counts",
          "daily_counts"
        ],
        "title": "AlertSummaryResponse",
        "type": "object"
      },
...
      "DailyAlertCountResponse": {
        "description": "How many alerts of one kind were raised on one UTC day.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AlertKind"
          }
        },
        "required": [
          "day",
          "kind",
          "count"
        ],
        "title": "DailyAlertCountResponse",
        "type": "object"
      },
...
      "Disposition": {
        "description": "A reviewer's verdict on an alert.",
        "enum": [
          "unreviewed",
          "true_positive",
          "false_positive"
        ],
        "title": "Disposition",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Specter",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/owners/{owner_id}/alerts": {
      "get": {
        "description": "Lists the owner's alerts of both kinds, or of one, newest first.",
        "operationId": "list_all_alerts_owners__owner_id__alerts_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlertKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Disposition"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disposition"
            }
          },
          {
            "in": "query",
            "name": "created_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Since"
            }
          },
          {
            "in": "query",
            "name": "created_until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Until"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List All Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/identity-matches": {
      "get": {
        "description": "Lists the owner's identity match alerts, newest first.",
        "operationId": "list_identity_matches_owners__owner_id__alerts_identity_matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Disposition"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disposition"
            }
          },
          {
            "in": "query",
            "name": "created_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Since"
            }
          },
          {
            "in": "query",
            "name": "created_until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Until"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Identity Matches",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/rules": {
      "get": {
        "description": "Lists the owner's rule alerts, newest first.",
        "operationId": "list_rule_firings_owners__owner_id__alerts_rules_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Disposition"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disposition"
            }
          },
          {
            "in": "query",
            "name": "created_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Since"
            }
          },
          {
            "in": "query",
            "name": "created_until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Until"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Rule Firings",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/summary": {
      "get": {
        "description": "Counts the owner's alerts by kind and review state, and by UTC day.",
        "operationId": "summarize_owner_alerts_owners__owner_id__alerts_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "created_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Since"
            }
          },
          {
            "in": "query",
            "name": "created_until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Summarize Owner Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/{alert_id}": {
      "get": {
        "description": "Returns one of the owner's alerts.",
        "operationId": "read_alert_owners__owner_id__alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/{alert_id}/acknowledge": {
      "post": {
        "description": "Marks the alert as seen, without a verdict.",
        "operationId": "acknowledge_alert_owners__owner_id__alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Acknowledge Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/{alert_id}/resolve": {
      "post": {
        "description": "Records the reviewer's verdict, which also acknowledges the alert.",
        "operationId": "resolve_alert_owners__owner_id__alerts__alert_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/owners/{owner_id}/alerts/{alert_id}/snapshot": {
      "get": {
        "description": "Returns the frame the alert was raised on.\n\nRaises:\n    NotFoundError: The alert has no snapshot, or retention already removed it.",
        "operationId": "read_snapshot_owners__owner_id__alerts__alert_id__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Snapshot",
        "tags": [
          "alerts"
        ]
      }
    },
...
  }
}